+++ db/migrate/20241103190746_add_career_to_user.rb
@@ -1,0 +1,5 @@
+class AddCareerToUser < ActiveRecord::Migration[7.2]
+  def change
+    add_reference :users, :career, foreign_key: true
+  end
+end