+++ spec/requests/users_controller_spec.rb
@@ -1,14 +1,14 @@
 require 'rails_helper'
 
 RSpec.describe "UsersControllers", type: :request do
-  let!(:user) { FactoryBot.create(:user) }
+  let!(:user) { FactoryBot.create(:user, career: nil) }
   let!(:token) { JsonWebTokenService.encode(user_id: user.id) }
   describe "GET /profile" do
     it "returns the current user with correct attributes" do
-      get "/profile", headers: { "Authorization" => token }
+      get "/profile", headers: { 'Authorization': "Bearer #{token}" }
       expect(response).to have_http_status(200)
-      expect(JSON.parse(response.body).first.keys).to eq([ "id", "name", "email", "description", "image_url" ])
-      expect(JSON.parse(response.body).first["name"]).to eq(user.name)
+      expect(JSON.parse(response.body).keys).to eq([ "id", "name", "email", "description", "image_url", "career" ])
+      expect(JSON.parse(response.body)["name"]).to eq(user.name)
     end
 
     it "returns an error if the token is invalid" do
@@ -22,6 +22,11 @@ RSpec.describe "UsersControllers", type: :request do
       expect(response).to have_http_status(400)
       expect(JSON.parse(response.body)).to eq({ "error" => I18n.t("error.sessions.missing_header") })
     end
+
+    it "the career is nil if the user does not have a career" do
+      get "/profile", headers: { 'Authorization': "Bearer #{token}" }
+      expect(JSON.parse(response.body)["career"]).to be_nil
+    end
   end
 
   describe "PUT /update" do
@@ -32,9 +37,49 @@ RSpec.describe "UsersControllers", type: :request do
         non_permited_attribute: "non-editable",
         email: "non-editable" }
       },
-      headers: { "Authorization" => token }
+      headers: { 'Authorization': "Bearer #{token}" }
       expect(response).to have_http_status(200)
       expect(JSON.parse(response.body)).to eq({ "message" => I18n.t("success.users.updated") })
+      email = user.email
+      user.reload
+      expect(user.name).to eq("Juan Pablo II")
+      expect(user.description).to eq("new description")
+      expect(user.email).to eq(email)
+    end
+
+    it "updating only career returns success" do
+      career = FactoryBot.create(:career)
+      put "/profile", params: { user: { career_id: career.id } }, headers: { 'Authorization': "Bearer #{token}" }
+      expect(response).to have_http_status(200)
+      expect(JSON.parse(response.body)).to eq({ "message" => I18n.t("success.users.updated") })
+      user.reload
+      expect(user.career.id).to eq(career.id)
+      expect(user.career.name).to eq(career.name)
+    end
+
+    it "updating only name returns success" do
+      put "/profile", params: { user: { name: "Juan Pablo II" } }, headers: { 'Authorization': "Bearer #{token}" }
+      expect(response).to have_http_status(200)
+      expect(JSON.parse(response.body)).to eq({ "message" => I18n.t("success.users.updated") })
+      user.reload
+      expect(user.name).to eq("Juan Pablo II")
+    end
+
+    it "updating only description returns success" do
+      put "/profile", params: { user: { description: "new description" } },
+      headers: { 'Authorization': "Bearer #{token}" }
+      expect(response).to have_http_status(200)
+      expect(JSON.parse(response.body)).to eq({ "message" => I18n.t("success.users.updated") })
+      user.reload
+      expect(user.description).to eq("new description")
+    end
+
+    it "returns an error if the career id is invalid" do
+      put "/profile",
+        params: { user: { career_id: 0 } },
+        headers: { 'Authorization': "Bearer #{token}" }
+      expect(response).to have_http_status(404)
+      expect(JSON.parse(response.body)).to eq({ "error" => I18n.t("error.careers.not_found") })
     end
   end
 end